added some conditons to add and delete employee and user defined exceptions are added

+++ employee_manager/src/main/java/io/suvadeep/employee_manager/service/EmployeeService.java
@@ -1,5 +1,6 @@
 package io.suvadeep.employee_manager.service;
 
+import io.suvadeep.employee_manager.exception.EmployeeAlreadyExists;
 import io.suvadeep.employee_manager.exception.EmployeeNotFoundException;
 import io.suvadeep.employee_manager.model.Employee;
 import io.suvadeep.employee_manager.repository.EmployeeRepository;
@@ -7,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -18,7 +20,11 @@ public class EmployeeService {
         this.employeeRepository = employeeRepository;
     }
 
-    public Employee registerEmployee(Employee employee) {
+    public Employee registerEmployee(Employee employee) throws EmployeeAlreadyExists {
+        Optional<Employee> optionalEmployee = employeeRepository.findEmployeeByEmail(employee.getEmployeeEmail());
+        if (optionalEmployee.isPresent()) {
+            throw new EmployeeAlreadyExists("Employee with email: " + employee.getEmployeeEmail() + " already exists.");
+        }
         employee.setEmployeeRandomCode(UUID.randomUUID().toString());
         return employeeRepository.save(employee);
     }
@@ -35,7 +41,10 @@ public class EmployeeService {
         return employeeRepository.findById(employeeId).orElseThrow(() -> new EmployeeNotFoundException("Employee with id: " + employeeId + " does not exists."));
     }
 
-    public void deleteEmployee(Long employeeId) {
+    public void deleteEmployee(Long employeeId) throws EmployeeNotFoundException {
+        if (!employeeRepository.existsById(employeeId)) {
+            throw new EmployeeNotFoundException("Employee with id: " + employeeId + " does not exists.");
+        }
         employeeRepository.deleteById(employeeId);
     }
 }
